feat: docs routes use shadcn sidebar with back button #570
/docs now uses the same SidebarProvider/SidebarInset shell as the app
instead of the marketing site-header. The docs sidebar header gets a
"Back" button (history.back with / fallback), the content groups doc
pages by section using shadcn SidebarMenu primitives, and the footer
shares UserSidebarFooter with the app — extracted from app-sidebar.tsx
and extended with a "Sign in" state so anonymous users on /docs still
have a clear auth entry point.
