+++ src/components/docs/docs-app-sidebar.tsx
@@ -1,0 +1,48 @@
+import { DocsSidebar } from '@/components/docs/docs-sidebar';
+import {
+  Sidebar,
+  SidebarContent,
+  SidebarFooter,
+  SidebarHeader,
+  SidebarMenu,
+  SidebarMenuButton,
+  SidebarMenuItem,
+} from '@/components/ui/sidebar';
+import { UserSidebarFooter } from '@/components/layout/user-sidebar-footer';
+import { useRouter } from '@tanstack/react-router';
+import { ArrowLeft } from 'lucide-react';
+
+export function DocsAppSidebar() {
+  const router = useRouter();
+
+  const handleBack = () => {
+    if (window.history.length > 1) {
+      router.history.back();
+    } else {
+      void router.navigate({ to: '/' });
+    }
+  };
+
+  return (
+    <Sidebar collapsible="icon">
+      <SidebarHeader>
+        <SidebarMenu>
+          <SidebarMenuItem>
+            <SidebarMenuButton onClick={handleBack} tooltip="Back">
+              <ArrowLeft />
+              <span>Back</span>
+            </SidebarMenuButton>
+          </SidebarMenuItem>
+        </SidebarMenu>
+      </SidebarHeader>
+      <SidebarContent>
+        <DocsSidebar />
+      </SidebarContent>
+      <SidebarFooter>
+        <SidebarMenu>
+          <UserSidebarFooter />
+        </SidebarMenu>
+      </SidebarFooter>
+    </Sidebar>
+  );
+}
